Rename modules and enforce privacy of coordinates
